Split up main content script
